requirements-test.py3.txt
django-mptt~=0.11.0
mock~=3.0.0
pytest-django~=3.10.0
pytest-pythonpath~=0.7.0
pytest~=4.6.11
pytz==2020.1
six~=1.15.0
